Module aws_sdk_kinesisvideo::error

Modules

access_denied_exception

See AccessDeniedException

account_channel_limit_exceeded_exception

See AccountChannelLimitExceededException

account_stream_limit_exceeded_exception

See AccountStreamLimitExceededException

client_limit_exceeded_exception

See ClientLimitExceededException

device_stream_limit_exceeded_exception

See DeviceStreamLimitExceededException

invalid_argument_exception

See InvalidArgumentException

invalid_device_exception

See InvalidDeviceException

invalid_resource_format_exception

See InvalidResourceFormatException

not_authorized_exception

See NotAuthorizedException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

tags_per_resource_exceeded_limit_exception

See TagsPerResourceExceededLimitException

version_mismatch_exception

See VersionMismatchException

Structs

AccessDeniedException

You do not have required permissions to perform this operation.

AccountChannelLimitExceededException

You have reached the maximum limit of active signaling channels for this AWS account

AccountStreamLimitExceededException

The number of streams created for the account is too high.

ClientLimitExceededException

Kinesis Video Streams has throttled the request because you have exceeded the limit of

CreateSignalingChannelError

CreateStreamError

DeleteSignalingChannelError

DeleteStreamError

DescribeSignalingChannelError

DescribeStreamError

DeviceStreamLimitExceededException

Not implemented.

GetDataEndpointError

GetSignalingChannelEndpointError

InvalidArgumentException

The value for this input parameter is invalid.

InvalidDeviceException

Not implemented.

InvalidResourceFormatException

The format of the StreamARN is invalid.

ListSignalingChannelsError

ListStreamsError

ListTagsForResourceError

ListTagsForStreamError

NotAuthorizedException

The caller is not authorized to perform this operation.

ResourceInUseException

The signaling channel is currently not available for this operation.

ResourceNotFoundException

Amazon Kinesis Video Streams can't find the stream that you specified.

TagResourceError

TagStreamError

TagsPerResourceExceededLimitException

You have exceeded the limit of tags that you can associate with the resource.

UntagResourceError

UntagStreamError

UpdateDataRetentionError

UpdateSignalingChannelError

UpdateStreamError

VersionMismatchException

The stream version that you specified is not the latest version. To get the latest

Enums

CreateSignalingChannelErrorKind

CreateStreamErrorKind

DeleteSignalingChannelErrorKind

DeleteStreamErrorKind

DescribeSignalingChannelErrorKind

DescribeStreamErrorKind

GetDataEndpointErrorKind

GetSignalingChannelEndpointErrorKind

ListSignalingChannelsErrorKind

ListStreamsErrorKind

ListTagsForResourceErrorKind

ListTagsForStreamErrorKind

TagResourceErrorKind

TagStreamErrorKind

UntagResourceErrorKind

UntagStreamErrorKind

UpdateDataRetentionErrorKind

UpdateSignalingChannelErrorKind

UpdateStreamErrorKind