Module aws_sdk_kinesisvideo::model

Modules

channel_info

See ChannelInfo

channel_name_condition

See ChannelNameCondition

resource_endpoint_list_item

See ResourceEndpointListItem

single_master_channel_endpoint_configuration

See SingleMasterChannelEndpointConfiguration

single_master_configuration

See SingleMasterConfiguration

stream_info

See StreamInfo

stream_name_condition

See StreamNameCondition

tag

See Tag

Structs

ChannelInfo

A structure that encapsulates a signaling channel's metadata and properties.

ChannelNameCondition

An optional input parameter for the ListSignalingChannels API. When this

ResourceEndpointListItem

An object that describes the endpoint of the signaling channel returned by the

SingleMasterChannelEndpointConfiguration

An object that contains the endpoint configuration for the SINGLE_MASTER

SingleMasterConfiguration

A structure that contains the configuration for the SINGLE_MASTER channel

StreamInfo

An object describing a Kinesis video stream.

StreamNameCondition

Specifies the condition that streams must satisfy to be returned when you list

Tag

A key and value pair that is associated with the specified signaling channel.

Enums

ApiName

ChannelProtocol

ChannelRole

ChannelType

ComparisonOperator

Status

UpdateDataRetentionOperation