Module aws_sdk_kinesisvideo::operation

Structs

CreateSignalingChannel

Creates a signaling channel.

CreateStream

Creates a new Kinesis video stream.

DeleteSignalingChannel

Deletes a specified signaling channel. DeleteSignalingChannel is an

DeleteStream

Deletes a Kinesis video stream and the data contained in the stream.

DescribeSignalingChannel

Returns the most current information about the signaling channel. You must specify

DescribeStream

Returns the most current information about the specified stream. You must specify

GetDataEndpoint

Gets an endpoint for a specified stream for either reading or writing. Use this

GetSignalingChannelEndpoint

Provides an endpoint for the specified signaling channel to send and receive messages.

ListSignalingChannels

Returns an array of ChannelInfo objects. Each object describes a

ListStreams

Returns an array of StreamInfo objects. Each object describes a

ListTagsForResource

Returns a list of tags associated with the specified signaling channel.

ListTagsForStream

Returns a list of tags associated with the specified stream.

TagResource

Adds one or more tags to a signaling channel. A tag is a

TagStream

Adds one or more tags to a stream. A tag is a key-value pair

UntagResource

Removes one or more tags from a signaling channel. In the request, specify only a tag

UntagStream

Removes one or more tags from a stream. In the request, specify only a tag key or

UpdateDataRetention

Increases or decreases the stream's data retention period by the value that you

UpdateSignalingChannel

Updates the existing signaling channel. This is an asynchronous operation and takes

UpdateStream

Updates stream metadata, such as the device name and media type.